feat: api adapterをBlockとUserで分離しmoduleに登録
